test(microwave): ✅ add strategy testing glove
